Move the SocId class to a separate file
